Message test with testing library
